src/main/java/com/java/design/patterns/creational/builder/StudentBuilder.java
package com.java.design.patterns.creational.builder;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.NoArgsConstructor;

@AllArgsConstructor(access = AccessLevel.PRIVATE)
@NoArgsConstructor(access = AccessLevel.PUBLIC)
public class StudentBuilder {
    protected int id;
    protected String firstName;
    protected String lastName;

    public StudentBuilder id(int id) {
        this.id = id;
        return this;
    }

    public StudentBuilder firstName(String firstName) {
        this.firstName = firstName;
        return this;
    }

    public StudentBuilder lastName(String lastName) {
        this.lastName = lastName;
        return this;
    }

    public Student build() {
        return new Student(this);
    }
}
